+++ packages/cloud-cognitive/src/components/_Canary/Canary.js
@@ -1,0 +1,63 @@
+/**
+ * Copyright IBM Corp. 2020, 2021
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import cx from 'classnames';
+import { string } from 'prop-types';
+import React from 'react';
+import { CodeSnippet } from 'carbon-components-react';
+
+// load the package settings direct, because Canary is used by settings.js
+import pkg from '../../global/js/package-settings';
+
+const blockClass = `${pkg.prefix}-canary`;
+
+/**
+ *  Canary component used when the component requested is not yet production
+ */
+export const Canary = (
+  { componentName, className, ...rest } /*, originalArgs*/
+) => {
+  const instructions = `
+import { pkg } from '@carbon/ibm-cloud-cognitive';
+// NOTE: must happen before component import
+pkg.component.${componentName} = true;
+`;
+  return (
+    <div className={cx(blockClass, className)} {...rest}>
+      <h2>
+        This component <strong>{componentName}</strong> is not ready yet.
+      </h2>
+      <p>
+        To enable this initialize package flags before any components are
+        loaded, passing an override object.
+      </p>
+      <br />
+      <p>e.g. in main.js</p>
+      <CodeSnippet type="multi">{instructions}</CodeSnippet>
+      <br />
+      <p>
+        View a live example on{' '}
+        <a href="https://codesandbox.io/s/example-component-olif5?file=/src/config.js">
+          codesadnbox
+        </a>
+        .
+      </p>
+    </div>
+  );
+};
+
+Canary.propTypes = {
+  /** Provide an optional class to be applied to the containing node */
+  className: string,
+
+  /** Name of the component that is not ready yet */
+  componentName: string.isRequired,
+};
+
+Canary.defaultProps = {
+  className: null,
+};